+++ odoo/local-src/specific_print/models/stock.py
@@ -90,8 +90,11 @@ class StockPicking(models.Model):
     @api.multi
     def print_products_label(self, printer=False):
         self.ensure_one()
-        if self.pack_operation_ids:
-            self.pack_operation_ids.print_product_label(printer=printer)
+
+        packs_to_print = self.pack_operation_ids.filtered(
+            lambda pack_op: not pack_op.product_id.is_do_not_print_label)
+        if packs_to_print:
+            packs_to_print.print_product_label(printer=printer)
 
     @api.multi
     def print_packages_label(self, quantity=1, printer=False):