+++ HISTORY.rst
@@ -8,6 +8,8 @@ Unreleased
 
 **Features and Improvements**
 
+* ALCYN-1695: Add a flag on product to not print label for these product
+
 **Bugfixes**
 
 **Build**